requirements.txt
Pypubsub==4.0.3
pyserial==3.5
python-can==4.0.0
pygame
socketcan
numpy